+++ facebook_auth/example/lib/main.dart
@@ -58,50 +58,35 @@ class _MyAppState extends State<AuthExample> {
   }
 
   Future<void> _login() async {
-    try {
-      // show a circular progress indicator
-      setState(() {
-        _checking = true;
-      });
-      _accessToken = await FacebookAuth.instance.login(); // by the fault we request the email and the public profile
+    // show a circular progress indicator
+    setState(() {
+      _checking = true;
+    });
+    final result = await FacebookAuth.instance.login(); // by the fault we request the email and the public profile
 
-      // loginBehavior is only supported for Android devices, for ios it will be ignored
-      // _accessToken = await FacebookAuth.instance.login(
-      //   permissions: ['email', 'public_profile', 'user_birthday', 'user_friends', 'user_gender', 'user_link'],
-      //   loginBehavior:
-      //       LoginBehavior.DIALOG_ONLY, // (only android) show an authentication dialog instead of redirecting to facebook app
-      // );
+    // loginBehavior is only supported for Android devices, for ios it will be ignored
+    // final result = await FacebookAuth.instance.login(
+    //   permissions: ['email', 'public_profile', 'user_birthday', 'user_friends', 'user_gender', 'user_link'],
+    //   loginBehavior: LoginBehavior
+    //       .DIALOG_ONLY, // (only android) show an authentication dialog instead of redirecting to facebook app
+    // );
+
+    if (result.status == LoginStatus.success) {
+      _accessToken = result.accessToken;
       _printCredentials();
       // get the user data
       // by default we get the userId, email,name and picture
       final userData = await FacebookAuth.instance.getUserData();
       // final userData = await FacebookAuth.instance.getUserData(fields: "email,birthday,friends,gender,link");
       _userData = userData;
-    } on FacebookAuthException catch (e) {
-      // if the facebook login fails
-      print(e.message); // print the error message in console
-      // check the error type
-      switch (e.errorCode) {
-        case FacebookAuthErrorCode.OPERATION_IN_PROGRESS:
-          print("You have a previous login operation in progress");
-          break;
-        case FacebookAuthErrorCode.CANCELLED:
-          print("login cancelled");
-          break;
-        case FacebookAuthErrorCode.FAILED:
-          print("login failed");
-          break;
-      }
-    } catch (e, s) {
-      // print in the logs the unknown errors
-      print(e);
-      print(s);
-    } finally {
-      // update the view
-      setState(() {
-        _checking = false;
-      });
+    } else {
+      print(result.status);
+      print(result.message);
     }
+
+    setState(() {
+      _checking = false;
+    });
   }
 
   Future<void> _logOut() async {