+++ facebook_auth_platform_interface/lib/src/method_cahnnel.dart
@@ -1,11 +1,11 @@
 import 'dart:convert';
 import 'package:flutter/services.dart';
+import 'login_result.dart';
 import '../flutter_facebook_auth_platform_interface.dart';
 import 'access_token.dart';
 import 'dart:io' show Platform;
 import 'package:flutter/foundation.dart' show kIsWeb, visibleForTesting;
 import 'login_behavior.dart';
-import 'facebook_auth_exception.dart';
 
 /// class to make calls to the facebook login SDK
 class MethodCahnnelFacebookAuth extends FacebookAuthPlatform {
@@ -20,7 +20,7 @@ class MethodCahnnelFacebookAuth extends FacebookAuthPlatform {
   /// [loginBehavior] (only Android) use this param to set the UI for the authentication,
   /// like webview, native app, or a dialog.
   @override
-  Future<AccessToken> login({
+  Future<LoginResult> login({
     List<String> permissions = const ['email', 'public_profile'],
     String loginBehavior = LoginBehavior.DIALOG_ONLY,
   }) async {
@@ -29,9 +29,10 @@ class MethodCahnnelFacebookAuth extends FacebookAuthPlatform {
         "permissions": permissions,
         "loginBehavior": loginBehavior,
       });
-      return AccessToken.fromJson(Map<String, dynamic>.from(result));
+      final token = AccessToken.fromJson(Map<String, dynamic>.from(result));
+      return LoginResult(status: LoginStatus.success, accessToken: token);
     } on PlatformException catch (e) {
-      throw FacebookAuthException(e.code, e.message);
+      return LoginResult.getResultFromException(e);
     }
   }
 
@@ -43,17 +44,19 @@ class MethodCahnnelFacebookAuth extends FacebookAuthPlatform {
   /// first add the following code to the queries element in your /app/manifest/AndroidManifest.xml file.
   /// For more info go to https://developers.facebook.com/docs/facebook-login/android
   @override
-  Future<AccessToken?> expressLogin() async {
+  Future<LoginResult> expressLogin() async {
     final tesing = Platform.environment.containsKey('FLUTTER_TEST');
     if (Platform.isAndroid || tesing) {
       try {
         final result = await channel.invokeMethod("expressLogin");
-        return AccessToken.fromJson(Map<String, dynamic>.from(result));
+        final token = AccessToken.fromJson(Map<String, dynamic>.from(result));
+        return LoginResult(status: LoginStatus.success, accessToken: token);
       } on PlatformException catch (e) {
-        throw FacebookAuthException(e.code, e.message);
+        return LoginResult.getResultFromException(e);
       }
     }
-    return null;
+    return LoginResult(
+        status: LoginStatus.failed, message: 'Method only allowed on Android');
   }
 
   /// retrive the user information using the GraphAPI
@@ -63,19 +66,15 @@ class MethodCahnnelFacebookAuth extends FacebookAuthPlatform {
   Future<Map<String, dynamic>> getUserData({
     String fields = "name,email,picture.width(200)",
   }) async {
-    try {
-      final result = await channel.invokeMethod("getUserData", {
-        "fields": fields,
-      });
-      if (kIsWeb) {
-        return Map<String, dynamic>.from(result);
-      } else {
-        return Platform.isAndroid
-            ? jsonDecode(result)
-            : Map<String, dynamic>.from(result); //null  or dynamic data
-      }
-    } on PlatformException catch (e) {
-      throw FacebookAuthException(e.code, e.message);
+    final result = await channel.invokeMethod("getUserData", {
+      "fields": fields,
+    });
+    if (kIsWeb) {
+      return Map<String, dynamic>.from(result);
+    } else {
+      return Platform.isAndroid
+          ? jsonDecode(result)
+          : Map<String, dynamic>.from(result); //null  or dynamic data
     }
   }